Move appose demo classes to separate appose package
