+++ src/test/java/org/mastodon/mamut/appose/ApposeNumpy1Example.java
@@ -1,0 +1,49 @@
+package org.mastodon.mamut.appose;
+
+import java.io.BufferedWriter;
+import java.io.File;
+import java.io.FileWriter;
+import java.io.IOException;
+import java.nio.file.Files;
+
+import org.apposed.appose.Appose;
+import org.apposed.appose.Environment;
+import org.apposed.appose.Service;
+
+public class ApposeNumpy1Example
+{
+	public static void main( String[] args ) throws IOException, InterruptedException
+	{
+		File envFile = Files.createTempFile( "env", "yml" ).toFile();
+		try (BufferedWriter writer = new BufferedWriter( new FileWriter( envFile ) ))
+		{
+			writer.write( "name: numpy1\n" );
+			writer.write( "channels:\n" );
+			writer.write( "  - conda-forge\n" );
+			writer.write( "dependencies:\n" );
+			writer.write( "  - python=3.10\n" );
+			writer.write( "  - pip\n" );
+			writer.write( "  - pip:\n" );
+			writer.write( "    - appose\n" );
+			writer.write( "  - numpy<1.24\n" );
+		}
+		envFile.deleteOnExit();
+		Environment env = Appose.file( envFile, "environment.yml" ).logDebug().build();
+		System.out.println( "Created environment" );
+
+		String script = "import numpy\n";
+
+		try (Service python = env.python())
+		{
+			Service.Task task = python.task( script );
+			task.waitFor();
+
+			// Verify that it worked.
+			if ( task.status != Service.TaskStatus.COMPLETE )
+			{
+				throw new RuntimeException( "Python script failed with error: " + task.error );
+			}
+			System.out.println( "Python script completed" );
+		}
+	}
+}
